Add files via upload
The dataset seems to be a sample from the the Consumer Financial Protection Bureau (CFPB) public complaint database. Business Goal: 
Find data-driven insights to reduce the resolution time for customer complaints to improve customer satisfaction and retention.
• KPIs:
○ Average resolution time per complaint
○ Percentage of complaints resolved within a certain timeframe (e.g., 3 days, 7 days)
Prepare dataset for the Machine Learning Model Prediction:
○ Build a predictive model to forecast the resolution time for new complaints based on historical data such as product, issue type, and company response (out of the scope).

Data cleaning process report:
The cleaning initially started defining the business goal CCRO this allow to design a strategy to cutting off data.
CCRO selection allowed the dropping off 4 columns (the red colored in table before).
A step for removing duplicates was performed.
A step of populating ‘State’ from non ‘NaN’ so we do not drop useful data in the later step of removing rows if NaN values are present. Here a factores dictionary of mapping ZCodes to states helped to run a script for populating ‘States’ whenever a zip code provide this information rescuing 13K row of useful data.
A step for removing additional columns was performed considering the proportion of NaN and when there was not possibility to populate (with fact meaning) those missing values.  This step dropped 4 additional columns (in purple here). 
A step of removing NaN to each row that had a NaN value was performed.
A step for changing column names was conducted.
A sampling step was conducted. At this point, df had 3.2M rows. To make more easy and efficient the later data processing it was necessary to reduce that number to no less than 10K rows. As a buffer I set it to be 50K rows in case more cleaning would be necessary. To achieve this a column analysis was performed to check which column was best to extract the 50K sample by an stratification method.
CCRO and df data in ‘issue’ and ‘product’ helped to weight ‘Issue' was best suited. Both columns are still important for stratification, with 'Issue' being slightly favored due to its granularity. Adjustments to the weights can be made based on additional insights or specific business requirements. For this step I will use ZipCode as a joint to merge datasets. I used the IRS dataset: SOI Tax Stats - Individual Income Tax Statistics - 2020 ZIP Code Data (SOI); this is representative of our cleaned dates because it is heavily representing years 2022 and 2023 account for near 42% or the universe of the 50K registries, 2020-2023 adds up to 62% ). IRS data was cleaned and worked on to get calculated columns to add categories of income rank and inequality rank by zip code
Complementary dataset: irs_zip_income_inequelity_cleaned5.csv (here to in this repository.